Removed boss-specific AP rewards from (Unstable) Abyssal Splinter due to being killable by only one player. Also, the retail AP reward is only about 200 AP instead of 37,500 AP per boss. Additionally, refactored related classes and methods.

+++ game-server/data/handlers/instance/abyss/UnstableSplinterpathInstance.java
@@ -22,7 +22,7 @@ import com.aionemu.gameserver.world.WorldMapInstance;
 public class UnstableSplinterpathInstance extends GeneralInstanceHandler {
 
 	private int destroyedFragments;
-	private int killedPazuzuWorms = 0;
+	private int killedPazuzuWorms;
 
 	public UnstableSplinterpathInstance(WorldMapInstance instance) {
 		super(instance);
@@ -33,23 +33,19 @@ public class UnstableSplinterpathInstance extends GeneralInstanceHandler {
 		super.onDie(npc);
 		final int npcId = npc.getNpcId();
 		switch (npcId) {
-			case 219554: // (ex 219942) Pazuzu the Life Current
-				spawnPazuzuHugeAetherFragment();
-				spawnPazuzuGenesisTreasureBoxes();
-				spawnPazuzuAbyssalTreasureBox();
-				spawnPazuzusTreasureBox();
+			case 219554: // Unstable Pazuzu
+				spawnPazuzuFragment();
+				spawnPazuzuTreasureBoxes();
 				break;
-			case 219553: // (ex 219941) Kaluva the Fourth Fragment
-				spawnKaluvaHugeAetherFragment();
-				spawnKaluvaGenesisTreasureBoxes();
-				spawnKaluvaAbyssalTreasureBox();
+			case 219553: // Unstable Kaluva the Fourth
+				spawnKaluvaFragment();
+				spawnKaluvaTreasureBoxes();
 				break;
-			case 219551: // (ex 219939) Unstable rukril
-			case 219552: // (ex 219940) Unstable ebonsoul
+			case 219551: // Unstable Rukril
+			case 219552: // Unstable Ebonsoul
 				if (getNpc(npcId == 219552 ? 219551 : 219552) == null) {
-					spawnDayshadeAetherFragment();
-					spawnDayshadeGenesisTreasureBoxes();
-					spawnDayshadeAbyssalTreasureChest();
+					spawnDayshadeFragment();
+					spawnDayshadeTreasureBoxes();
 				} else {
 					sendMsg(npcId == 219551 ? SM_SYSTEM_MESSAGE.STR_MSG_IDAbRe_Core_NmdC_Light_Die() : SM_SYSTEM_MESSAGE.STR_MSG_IDAbRe_Core_NmdC_Dark_Die());
 					ThreadPoolManager.getInstance().schedule(() -> {
@@ -63,44 +59,37 @@ public class UnstableSplinterpathInstance extends GeneralInstanceHandler {
 				}
 				npc.getController().delete();
 				break;
-			case 283204: // ex 284022
+			case 283204: // Piece of Splendor
 				Npc ebonsoul = getNpc(219552);
-				if (ebonsoul != null && !ebonsoul.isDead()) {
-					if (PositionUtil.isInRange(npc, ebonsoul, 5)) {
-						ebonsoul.getEffectController().removeEffect(19159);
-						deleteAliveNpcs(281907);
-						break;
-					}
+				if (ebonsoul != null && !ebonsoul.isDead() && PositionUtil.isInRange(npc, ebonsoul, 5)) {
+					ebonsoul.getEffectController().removeEffect(19159);
+					deleteAliveNpcs(281907);
+					break;
 				}
 				npc.getController().delete();
 				break;
-			case 283205: // ex 284023:
+			case 283205: // Piece of Midnight
 				Npc rukril = getNpc(219551);
-				if (rukril != null && !rukril.isDead()) {
-					if (PositionUtil.isInRange(npc, rukril, 5)) {
-						rukril.getEffectController().removeEffect(19266);
-						deleteAliveNpcs(281908);
-						break;
-					}
+				if (rukril != null && !rukril.isDead() && PositionUtil.isInRange(npc, rukril, 5)) {
+					rukril.getEffectController().removeEffect(19266);
+					deleteAliveNpcs(281908);
+					break;
 				}
 				npc.getController().delete();
 				break;
-			case 219563: // ex 219951 unstable Yamennes Painflare
-			case 219555: // ex 219943 strengthened Yamennes Blindsight
-				spawnYamennesGenesisTreasureBoxes();
-				spawnYamennesAbyssalTreasureBox(npcId == 219563 ? 701579 : 701580);
-				deleteAliveNpcs(219586); // Ex 219974
+			case 219563: // Unstable Yamennes Painflare
+			case 219555: // Durable Yamennes Blindsight
+				spawnYamennesTreasureBoxes(npcId == 219563 ? 701579 : 701580);
+				deleteAliveNpcs(219586); // Summoned Unstable Ametgolem
 				spawn(730317, 328.476f, 762.585f, 197.479f, (byte) 90); // Exit
-				for (Player p : instance.getPlayersInside())
-					SkillEngine.getInstance().applyEffectDirectly(19283, p, p);
+				instance.forEachPlayer(p -> SkillEngine.getInstance().applyEffectDirectly(19283, p, p));
 				break;
-			case 701588: // HugeAetherFragment
+			case 701588: // Huge Aether Fragment
 				destroyedFragments++;
 				onFragmentKill();
 				npc.getController().delete();
 				break;
-
-			case 283206: // ex 284024:
+			case 283206: // Luminous Waterworm
 				if (++killedPazuzuWorms == 4) {
 					killedPazuzuWorms = 0;
 					Npc pazuzu = getNpc(219554);
@@ -109,13 +98,11 @@ public class UnstableSplinterpathInstance extends GeneralInstanceHandler {
 						pazuzu.getEffectController().removeEffect(19291);
 					}
 				}
-				npc.getController().delete();
 				break;
-			case 219567: // ex 219955
-			case 219579: // ex 219967
-			case 219580: // ex 219968 Spawn Gate
-				removeSummoned();
-				npc.getController().delete();
+			case 219567: // Spawn Gate
+			case 219579: // Spawn Gate
+			case 219580: // Spawn Gate
+				deleteSummons();
 				break;
 		}
 	}
@@ -127,17 +114,15 @@ public class UnstableSplinterpathInstance extends GeneralInstanceHandler {
 
 	@Override
 	public void handleUseItemFinish(Player player, Npc npc) {
-		switch (npc.getNpcId()) {
-			case 700957:
-				sendMsg(SM_SYSTEM_MESSAGE.STR_MSG_IDAbRe_Core_NmdDH_Wakeup());
-				spawn(219563, 329.70886f, 733.8744f, 197.60938f, (byte) 0);
-				npc.getController().delete();
-				break;
-			case 701589:
-				sendMsg(SM_SYSTEM_MESSAGE.STR_MSG_IDAbRe_Core_NmdD_Wakeup());
-				spawn(219555, 329.70886f, 733.8744f, 197.60938f, (byte) 0);
-				npc.getController().delete();
-				break;
+		int bossId = switch (npc.getNpcId()) { // Artifact of Protection
+			case 700957 -> 219563;
+			case 701589 -> 219555;
+			default -> 0;
+		};
+		if (bossId != 0) {
+			sendMsg(SM_SYSTEM_MESSAGE.STR_MSG_IDAbRe_Core_NmdDH_Wakeup());
+			spawn(bossId, 329.70886f, 733.8744f, 197.60938f, (byte) 0);
+			npc.getController().delete();
 		}
 	}
 
@@ -147,69 +132,55 @@ public class UnstableSplinterpathInstance extends GeneralInstanceHandler {
 		return true;
 	}
 
-	private void spawnPazuzuHugeAetherFragment() {
+	private void spawnPazuzuFragment() {
 		spawn(701588, 669.576f, 335.135f, 465.895f, (byte) 0);
 	}
 
-	private void spawnPazuzuGenesisTreasureBoxes() {
-		spawn(701576, 651.53204f, 357.085f, 466.1315f, (byte) 66);
-		spawn(701576, 647.00446f, 357.2484f, 465.8960f, (byte) 0);
-		spawn(701576, 653.8384f, 360.39508f, 466.4391f, (byte) 100);
+	private void spawnPazuzuTreasureBoxes() {
+		sendMsg(SM_SYSTEM_MESSAGE.STR_MSG_IDAbRe_Core_NmdC_BoxSpawn());
+		spawn(701575, 649.24286f, 361.33755f, 466.0427f, (byte) 33); // Abyssal Treasure Box
+		spawn(701576, 651.53204f, 357.085f, 466.1315f, (byte) 66); // Genesis Treasure Box
+		spawn(701576, 647.00446f, 357.2484f, 465.8960f, (byte) 0); // Genesis Treasure Box
+		spawn(701576, 653.8384f, 360.39508f, 466.4391f, (byte) 100); // Genesis Treasure Box
+		if (Rnd.chance() < 20) // 20% chance, not retail
+			spawn(700861, 649.243f, 362.338f, 466.0118f, (byte) 0); // Pazuzu's Treasure Box
 	}
 
-	private void spawnPazuzuAbyssalTreasureBox() {
-		spawn(701575, 649.24286f, 361.33755f, 466.0427f, (byte) 33);
-	}
-
-	private void spawnPazuzusTreasureBox() {
-		if (Rnd.chance() < 20) { // 20% chance, not retail
-			spawn(700861, 649.243f, 362.338f, 466.0118f, (byte) 0);
-		}
-	}
-
-	private void spawnKaluvaHugeAetherFragment() {
+	private void spawnKaluvaFragment() {
 		spawn(701588, 633.7498f, 557.8822f, 424.99347f, (byte) 6);
 	}
 
-	private void spawnKaluvaGenesisTreasureBoxes() {
-		spawn(701576, 601.2931f, 584.66705f, 422.9955f, (byte) 6);
-		spawn(701576, 597.2156f, 583.95416f, 423.3474f, (byte) 66);
-		spawn(701576, 602.9586f, 589.2678f, 422.8296f, (byte) 100);
+	private void spawnKaluvaTreasureBoxes() {
+		sendMsg(SM_SYSTEM_MESSAGE.STR_MSG_IDAbRe_Core_NmdC_BoxSpawn());
+		spawn(701576, 601.2931f, 584.66705f, 422.9955f, (byte) 6); // Genesis Treasure Box
+		spawn(701576, 597.2156f, 583.95416f, 423.3474f, (byte) 66); // Genesis Treasure Box
+		spawn(701576, 602.9586f, 589.2678f, 422.8296f, (byte) 100); // Genesis Treasure Box
+		spawn(701577, 598.82776f, 588.25946f, 422.7739f, (byte) 113); // Abyssal Treasure Box
 	}
 
-	private void spawnKaluvaAbyssalTreasureBox() {
-		spawn(701577, 598.82776f, 588.25946f, 422.7739f, (byte) 113);
-	}
-
-	private void spawnDayshadeAetherFragment() {
+	private void spawnDayshadeFragment() {
 		spawn(701588, 452.89706f, 692.36084f, 433.96838f, (byte) 6);
 	}
 
-	private void spawnDayshadeGenesisTreasureBoxes() {
-		spawn(701576, 408.10938f, 650.9015f, 439.28332f, (byte) 66);
-		spawn(701576, 402.40375f, 655.55237f, 439.26288f, (byte) 33);
-		spawn(701576, 406.74445f, 655.5914f, 439.2548f, (byte) 100);
-	}
-
-	private void spawnDayshadeAbyssalTreasureChest() {
+	private void spawnDayshadeTreasureBoxes() {
 		sendMsg(SM_SYSTEM_MESSAGE.STR_MSG_IDAbRe_Core_NmdC_BoxSpawn());
-		spawn(701578, 404.891f, 650.2943f, 439.2548f, (byte) 130);
+		spawn(701576, 408.10938f, 650.9015f, 439.28332f, (byte) 66); // Genesis Treasure Box
+		spawn(701576, 402.40375f, 655.55237f, 439.26288f, (byte) 33); // Genesis Treasure Box
+		spawn(701576, 406.74445f, 655.5914f, 439.2548f, (byte) 100); // Genesis Treasure Box
+		spawn(701578, 404.891f, 650.2943f, 439.2548f, (byte) 130); // Abyssal Treasure Box
 	}
 
-	private void spawnYamennesGenesisTreasureBoxes() {
-		spawn(701576, 326.978f, 729.8414f, 197.7078f, (byte) 16);
-		spawn(701576, 326.5296f, 735.13324f, 197.6681f, (byte) 66);
-		spawn(701576, 329.8462f, 738.41095f, 197.7329f, (byte) 3);
+	private void spawnYamennesTreasureBoxes(int npcId) {
+		sendMsg(SM_SYSTEM_MESSAGE.STR_MSG_IDAbRe_Core_NmdC_BoxSpawn());
+		spawn(701576, 326.978f, 729.8414f, 197.7078f, (byte) 16); // Genesis Treasure Box
+		spawn(701576, 326.5296f, 735.13324f, 197.6681f, (byte) 66); // Genesis Treasure Box
+		spawn(701576, 329.8462f, 738.41095f, 197.7329f, (byte) 3); // Genesis Treasure Box
+		spawn(npcId, 330.891f, 733.2943f, 197.6404f, (byte) 113); // Abyssal Treasure Box
 	}
 
-	private void spawnYamennesAbyssalTreasureBox(int npcId) {
-		spawn(npcId, 330.891f, 733.2943f, 197.6404f, (byte) 113);
-	}
-
-	private void removeSummoned() {
-		if (instance.getNpcs(219567, 219579, 219580).stream().allMatch(Creature::isDead)) {
+	private void deleteSummons() {
+		if (instance.getNpcs(219567, 219579, 219580).stream().allMatch(Creature::isDead))
 			deleteAliveNpcs(219565, 219566); // Summoned Unstable Orkanimum, Summoned Unstable Lapilima
-		}
 	}
 
 	private void onFragmentKill() {
@@ -217,8 +188,8 @@ public class UnstableSplinterpathInstance extends GeneralInstanceHandler {
 			case 1 -> sendMsg(SM_SYSTEM_MESSAGE.STR_MSG_IDAbRe_Core_Artifact_Die_01());
 			case 2 -> sendMsg(SM_SYSTEM_MESSAGE.STR_MSG_IDAbRe_Core_Artifact_Die_02());
 			case 3 -> {
-				deleteAliveNpcs(701589);
-				spawn(700957, 326.1821f, 766.9640f, 202.1832f, (byte) 100, 79);
+				deleteAliveNpcs(701589); // Artifact of Protection (Hard Mode)
+				spawn(700957, 326.1821f, 766.9640f, 202.1832f, (byte) 100, 79); // Artifact of Protection (Easy Mode)
 				sendMsg(SM_SYSTEM_MESSAGE.STR_MSG_IDAbRe_Core_Artifact_Die_03());
 			}
 		}
